feat: add SpeedInsights and i18n-aware routing fixes
- import and mount Vercel SpeedInsights in app.vue inside ClientOnly
  to avoid SSR warnings and enable client-only performance checks.
- extract full HTML scaffold previously in layouts/default.vue into a new
  layouts/_oldone.vue for reference; keep default layout minimal to only
  render Header, BaseMain and BaseFooter.
- add i18n head handling and canonical URL computation in the new
  layouts/_oldone.vue (useLocaleHead, useRoute, useRequestURL) to support
  hreflang/seo meta when needed.
- update ArticleCard to use localePath for links so blog routes respect
  current locale and produce correct localized URLs.

These changes improve client-side performance tooling and fix locale-
aware routing and SEO head handling while keeping the default layout
lightweight.
